saving cam tensor as float32

+++ fake_cam_gen.py
@@ -22,6 +22,6 @@ for img_path in glob.glob("fake_cam/*.png"):
     img = cv2.imread(img_path, 0)
     keys = torch.tensor([0])
     cam = cv2.resize(img, (125, 94))
-    cam = torch.from_numpy(cam[np.newaxis, :, :])
+    cam = torch.from_numpy(cam[np.newaxis, :, :].astype("float32"))
     cam_npy = {'keys': keys, 'cam': cam, 'high_res': cv2.resize(img, (500, 375))[np.newaxis, :, :]}
     np.save(f"result/cam/{osp.basename(img_path)[:-4]}.npy", cam_npy)
